+++ src/main/java/gov/usgs/volcanoes/swarm/chooser/WWSPanel.java
@@ -34,6 +34,7 @@ public class WWSPanel extends DataSourcePanel {
   }
 
   /**
+   * Reset source.
    * @see gov.usgs.volcanoes.swarm.chooser.DataSourcePanel#resetSource(java.lang.String)
    */
   public void resetSource(String source) {
@@ -58,10 +59,11 @@ public class WWSPanel extends DataSourcePanel {
   @Override
   protected void createPanel() {
     createFields();
-    FormLayout layout = new FormLayout("right:max(20dlu;pref), 3dlu, 40dlu, 0dlu, 126dlu", "");
+    FormLayout layout = new FormLayout("right:max(20dlu;pref), 3dlu, 40dlu, 0dlu, 156dlu", "");
 
     DefaultFormBuilder builder = new DefaultFormBuilder(layout).border(Borders.DIALOG);
-    builder.append(new JLabel("Use this data source to connect to a Winston Wave Server (WWS)."),
+    builder.append(new JLabel(
+        "Use this data source to connect to a Winston Wave Server (WWS) or OSOP WS (OWS)."),
         5);
     builder.nextLine();
     builder.appendSeparator();
@@ -70,7 +72,7 @@ public class WWSPanel extends DataSourcePanel {
     builder.nextLine();
     builder.append("Port:");
     builder.append(wwsPort);
-    builder.append(" Winston default: 16022");
+    builder.append(" Winston default: 16022, OWS default: 16032");
     builder.nextLine();
 
     builder.append("Timeout:");